+++ src/shadowcasting.h
@@ -81,7 +81,7 @@ struct four_quadrants {
 // We merge all of the absorption values by taking their cumulative average.
 inline float sight_calc( const float &numerator, const float &transparency, const int &distance )
 {
-    return numerator / static_cast<float>( exp( transparency * distance ) );
+    return numerator / std::exp( transparency * distance );
 }
 inline bool sight_check( const float &transparency, const float &/*intensity*/ )
 {
